// data/github/current/ui/packages/memex/src/client/api/side-panel/contracts.ts
export interface IssueMetadataDescription {
  body: string | null
  bodyHtml: string | null
  editedAt?: string
}
// ...
export type Reactions = {[value in ReactionEmotion]?: Array<string>}

export const ReactionEmotion = {
  HEART: 'heart',
  TADA: 'tada',
  SMILE: 'smile',
  ROCKET: 'rocket',
  EYES: 'eyes',
  THUMBS_UP: '+1',
  THUMBS_DOWN: '-1',
  THINKING_FACE: 'thinking_face',
} as const
export type ReactionEmotion = ObjectValues<typeof ReactionEmotion>

// More information can be found at https://docs.github.com/graphql/reference/enums#commentauthorassociation
export const CommentAuthorAssociation = {
  COLLABORATOR: 'COLLABORATOR',
  CONTRIBUTOR: 'CONTRIBUTOR',
  FIRST_TIMER: 'FIRST_TIMER',
  FIRST_TIME_CONTRIBUTOR: 'FIRST_TIME_CONTRIBUTOR',
  MANNEQUIN: 'MANNEQUIN',
  MEMBER: 'MEMBER',
  NONE: 'NONE',
  OWNER: 'OWNER',
} as const
export type CommentAuthorAssociation = ObjectValues<typeof CommentAuthorAssociation>